# Optimize multiplication operations