This also fixes a broken scaling with `Normed`.
```julia
julia> 0.6N0f8 * RGB{N0f16}(0.2)
RGB{N0f16}(0.00047,0.00047,0.00047) # master
RGB{N0f16}(0.12,0.12,0.12)          # this PR
```

Note that the result of the `N0f16` operation will be slightly different due to rounding errors.
```julia
julia> 0.72992N0f16 * Gray(0.72992N0f16)
Gray{N0f16}(0.53278) # master
Gray{N0f16}(0.53277) # this PR
```
In some cases, `master` is theoretically more efficient, but the benchmarks showed little effect. For this reason, I simplified some specialized methods. Also, this does not optimize `⊗` for RGBs.

Fixes #166