@kimikage 2021-04-29
### Benchmarks

<details>
<summary>Script</summary>

```julia
using ColorVectorSpace, ColorTypes, FixedPointNumbers, BenchmarkTools

BenchmarkTools.DEFAULT_PARAMETERS.seconds = 10
N = 1000

g1_n0f8  = rand(Gray{N0f8}, N, N);
g2_n0f8  = rand(Gray{N0f8}, N, N);
g1_n0f16 = rand(Gray{N0f16}, N, N);
g2_n0f16 = rand(Gray{N0f16}, N, N);
g1_f32   = rand(Gray{Float32}, N, N);
g2_f32   = rand(Gray{Float32}, N, N);
g1_f64   = rand(Gray{Float64}, N, N);
g2_f64   = rand(Gray{Float64}, N, N);

rgb1_n0f8  = rand(RGB{N0f8}, N, N);
rgb2_n0f8  = rand(RGB{N0f8}, N, N);
rgb1_n0f16 = rand(RGB{N0f16}, N, N);
rgb2_n0f16 = rand(RGB{N0f16}, N, N);
rgb1_f32   = rand(RGB{Float32}, N, N);
rgb2_f32   = rand(RGB{Float32}, N, N);
rgb1_f64   = rand(RGB{Float64}, N, N);
rgb2_f64   = rand(RGB{Float64}, N, N);

for r in (0.5N0f8, 0.5, 1)
    @show r
    @btime $r .* $g1_n0f8;
    @btime $r .* $g1_n0f16;
    @btime $r .* $g1_f32;
    @btime $r .* $g1_f64;
end

for op in (.*, .⋅, .⊙ , .⊗)
    @show op
    @btime $op($g1_n0f8,  $g2_n0f8);
    @btime $op($g1_n0f16, $g2_n0f16);
    @btime $op($g1_f32,   $g2_f32);
    @btime $op($g1_f64,   $g2_f64);
end

for r in (0.5N0f8, 0.5, 1)
    @show r
    @btime $r .* $rgb1_n0f8;
    @btime $r .* $rgb1_n0f16;
    @btime $r .* $rgb1_f32;
    @btime $r .* $rgb1_f64;
end

for op in (.⋅, .⊙)
    @show op
    @btime $op($rgb1_n0f8,  $rgb2_n0f8);
    @btime $op($rgb1_n0f16, $rgb2_n0f16);
    @btime $op($rgb1_f32,   $rgb2_f32);
    @btime $op($rgb1_f64,   $rgb2_f64);
end
```
</details>

```julia
julia> versioninfo()
Julia Version 1.6.1
Commit 6aaedecc44 (2021-04-23 05:59 UTC)
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-11.0.1 (ORCJIT, skylake)
```

| operation            | before   | after    |                  |
|:---------------------|:--------:|:--------:|:----------------:|
|`0.5N0f8 .* g1_n0f8`  | 1.649 ms | 0.070 ms |:heavy_check_mark:|
|`0.5N0f8 .* g1_n0f16` | 2.011 ms | 0.300 ms |:heavy_check_mark:|
|`0.5N0f8 .* g1_f32`   | 0.574 ms | 0.596 ms |                  |
|`0.5N0f8 .* g1_f64`   | 1.513 ms | 1.475 ms |                  |
|`0.5 .* g1_n0f8`      | 1.196 ms | 1.216 ms |                  |
|`0.5 .* g1_n0f16`     | 1.233 ms | 1.320 ms |                  |
|`0.5 .* g1_f32`       | 1.263 ms | 1.294 ms |                  |
|`0.5 .* g1_f64`       | 1.485 ms | 1.412 ms |                  |
|`1 .* g1_n0f8`        | 0.555 ms | 0.581 ms |                  |
|`1 .* g1_n0f16`       | 0.589 ms | 0.598 ms |                  |
|`1 .* g1_f32`         | 0.624 ms | 0.612 ms |                  |
|`1 .* g1_f64`         | 1.433 ms | 1.496 ms |                  |
|`g1_n0f8  .* g2_n0f8` | 1.966 ms | 0.082 ms |:heavy_check_mark:|
|`g1_n0f16 .* g2_n0f16`| 2.381 ms | 0.329 ms |:heavy_check_mark:|
|`g1_f32   .* g2_f32`  | 0.764 ms | 0.765 ms |                  |
|`g1_f64   .* g2_f64`  | 1.708 ms | 1.800 ms |                  |
|`g1_n0f8  .⋅ g2_n0f8` | 1.966 ms | 0.083 ms |:heavy_check_mark:|
|`g1_n0f16 .⋅ g2_n0f16`| 2.408 ms | 0.381 ms |:heavy_check_mark:|
|`g1_f32   .⋅ g2_f32`  | 0.775 ms | 0.784 ms |                  |
|`g1_f64   .⋅ g2_f64`  | 1.789 ms | 1.785 ms |                  |
|`g1_n0f8  .⊙ g2_n0f8` | 1.966 ms | 0.088 ms |:heavy_check_mark:|
|`g1_n0f16 .⊙ g2_n0f16`| 2.364 ms | 0.379 ms |:heavy_check_mark:|
|`g1_f32   .⊙ g2_f32`  | 0.788 ms | 0.800 ms |                  |
|`g1_f64   .⊙ g2_f64`  | 1.810 ms | 1.823 ms |                  |
|`g1_n0f8  .⊗ g2_n0f8` | 1.987 ms | 0.083 ms |:heavy_check_mark:|
|`g1_n0f16 .⊗ g2_n0f16`| 2.393 ms | 0.344 ms |:heavy_check_mark:|
|`g1_f32   .⊗ g2_f32`  | 0.775 ms | 0.754 ms |                  |
|`g1_f64   .⊗ g2_f64`  | 1.767 ms | 1.786 ms |                  |

| operation                | before   | after    |                  |
|:-------------------------|:--------:|:--------:|:----------------:|
|`0.5N0f8 .* rgb1_n0f8`    | 4.148 ms | 0.645 ms |:heavy_check_mark:|
|`0.5N0f8 .* rgb1_n0f16`   |(4.442 ms)| 2.330 ms |:heavy_check_mark:|
|`0.5N0f8 .* rgb1_f32`     | 2.269 ms | 2.453 ms |                  |
|`0.5N0f8 .* rgb1_f64`     | 4.645 ms | 4.725 ms |                  |
|`0.5 .* rgb1_n0f8`        | 3.788 ms | 3.876 ms |                  |
|`0.5 .* rgb1_n0f16`       | 4.038 ms | 4.082 ms |                  |
|`0.5 .* rgb1_f32`         | 4.082 ms | 4.026 ms |                  |
|`0.5 .* rgb1_f64`         | 4.734 ms | 4.700 ms |                  |
|`1 .* rgb1_n0f8`          | 2.022 ms | 1.933 ms |                  |
|`1 .* rgb1_n0f16`         | 3.022 ms | 3.107 ms |                  |
|`1 .* rgb1_f32`           | 2.370 ms | 2.494 ms |                  |
|`1 .* rgb1_f64`           | 4.772 ms | 4.679 ms |                  |
|`rgb1_n0f8  .⋅ rgb2_n0f8` | 1.074 ms | 1.037 ms |                  |
|`rgb1_n0f16 .⋅ rgb2_n0f16`| 4.121 ms | 4.045 ms |                  |
|`rgb1_f32   .⋅ rgb2_f32`  | 1.512 ms | 1.490 ms |                  |
|`rgb1_f64   .⋅ rgb2_f64`  | 3.645 ms | 3.669 ms |                  |
|`rgb1_n0f8  .⊙ rgb2_n0f8` |10.662 ms | 0.866 ms |:heavy_check_mark:|
|`rgb1_n0f16 .⊙ rgb2_n0f16`|11.465 ms | 2.570 ms |:heavy_check_mark:|
|`rgb1_f32   .⊙ rgb2_f32`  | 3.019 ms | 3.062 ms |                  |
|`rgb1_f64   .⊙ rgb2_f64`  | 5.793 ms | 5.728 ms |                  |