@kimikage 2021-05-02
I am going to implement a similar modification for division, and fix #154.